+++ frontend_tests/node_tests/settings_org.js
@@ -165,7 +165,8 @@ function test_submit_settings_form(override, submit_form) {
         realm_bot_creation_policy: settings_bots.bot_creation_policy_values.restricted.code,
         realm_email_address_visibility:
             settings_config.email_address_visibility_values.admins_only.code,
-        realm_add_emoji_by_admins_only: true,
+        realm_add_custom_emoji_policy:
+            settings_config.add_custom_emoji_policy_values.by_admins_only.code,
         realm_create_stream_by_admins_only: true,
         realm_waiting_period_threshold: 1,
         realm_default_language: '"es"',
@@ -210,9 +211,10 @@ function test_submit_settings_form(override, submit_form) {
     create_stream_policy_elem.attr("id", "id_realm_create_stream_policy");
     create_stream_policy_elem.data = () => "number";
 
-    const add_emoji_by_admins_only_elem = $("#id_realm_add_emoji_by_admins_only");
-    add_emoji_by_admins_only_elem.val("by_anyone");
-    add_emoji_by_admins_only_elem.attr("id", "id_realm_add_emoji_by_admins_only");
+    const add_custom_emoji_policy_elem = $("#id_realm_add_custom_emoji_policy");
+    add_custom_emoji_policy_elem.val("1");
+    add_custom_emoji_policy_elem.attr("id", "id_realm_add_custom_emoji_policy");
+    add_custom_emoji_policy_elem.data = () => "number";
 
     const bot_creation_policy_elem = $("#id_realm_bot_creation_policy");
     bot_creation_policy_elem.val("1");
@@ -233,7 +235,7 @@ function test_submit_settings_form(override, submit_form) {
     subsection_elem.set_find_results(".prop-element", [
         bot_creation_policy_elem,
         email_address_visibility_elem,
-        add_emoji_by_admins_only_elem,
+        add_custom_emoji_policy_elem,
         create_stream_policy_elem,
         invite_to_stream_policy_elem,
     ]);
@@ -246,7 +248,7 @@ function test_submit_settings_form(override, submit_form) {
         bot_creation_policy: 1,
         invite_to_stream_policy: 1,
         email_address_visibility: 1,
-        add_emoji_by_admins_only: false,
+        add_custom_emoji_policy: 1,
         create_stream_policy: 2,
     };
     assert.deepEqual(data, expected_value);