+++ static/js/server_events_dispatch.js
@@ -176,7 +176,7 @@ export function dispatch_normal_event(event) {
 
         case "realm": {
             const realm_settings = {
-                add_emoji_by_admins_only: settings_emoji.update_custom_emoji_ui,
+                add_custom_emoji_policy: settings_emoji.update_custom_emoji_ui,
                 allow_edit_history: noop,
                 allow_message_deleting: noop,
                 allow_message_editing: noop,