+++ zerver/tests/test_events.py
@@ -2047,6 +2047,7 @@ class RealmPropertyActionTest(BaseAction):
             message_content_delete_limit_seconds=[1000, 1100, 1200],
             invite_to_realm_policy=[6, 4, 3, 2, 1],
             move_messages_between_streams_policy=[4, 3, 2, 1],
+            add_custom_emoji_policy=[2, 1],
         )
 
         vals = test_values.get(name)