+++ zerver/tests/test_home.py
@@ -129,7 +129,7 @@ class HomeTest(ZulipTestCase):
         "promote_sponsoring_zulip",
         "prompt_for_invites",
         "queue_id",
-        "realm_add_emoji_by_admins_only",
+        "realm_add_custom_emoji_policy",
         "realm_allow_edit_history",
         "realm_allow_message_deleting",
         "realm_allow_message_editing",